seperate VueScreener into section components
